Move some of the algebraic structures to felix format
Based a lot on what was done in the denotational-hardware, see:
* HasAlgebra
* Categorical.Raw.Monoid
* Categorical.Laws.Monoid
in:
https://github.com/conal/denotational-hardware/blob/main/HasAlgebra.agda

However I've tried to stick to stdlib as much as possible, therefore I
didn't port over `Has*' records and tried to use as much stdlib as
possible.  However I am not sure if that was a good call yet.